_minLength field should hvae type uint8_t

+++ include/sqids/sqids.hpp
@@ -134,7 +134,7 @@ private:
 
     std::string           _alphabet;
     std::set<std::string> _blocklist;
-    size_t                _minLength;
+    uint8_t               _minLength;
 };
 
 ///
